+++ src/Exception/Logic/ClientClosedException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Anktx\Kafka\Client\Exception\Logic;
+
+final class ClientClosedException extends LogicException
+{
+    public static function forMethod(string $method): self
+    {
+        return new self(\sprintf('Cannot call %s(): the client is closed', $method));
+    }
+}